Moved the message parsing into its own module.
